+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.29.0] - 2024-08-21
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Info about no recipe from author.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Divided `MealContainer` content into a few components.
+- [Marek Kurańda](https://github.com/mjkuranda): Some of API types to make it up to date.
+
 ## [1.28.0] - 2024-08-17
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `react-hook-form` library.